% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_empty.R
\name{gg_empty}
\alias{gg_empty}
\title{gg_empty}
\usage{
gg_empty(
  text = NULL,
  text.size = 12,
  title = NULL,
  title.size = 8,
  lib.path = NULL,
  safer_check = TRUE
)
}
\arguments{
\item{text}{character string of the message to display.}

\item{text.size}{numeric value of the text size (in points).}

\item{title}{character string of the graph title.}

\item{title.size}{numeric value of the title size (in points).}

\item{lib.path}{character vector specifying the absolute pathways of the directories containing the required packages if not in the default directories. Ignored if NULL.}

\item{safer_check}{Single logical value. Perform some "safer" checks (see https://github.com/safer-r)? If TRUE, checkings are performed before main code running: 1) R classical operators (like "<-") not overwritten by another package because of the R scope and 2) required functions and related packages effectively present in local R lybraries. Set to FALSE if this fonction is used inside another "safer" function to avoid pointless multiple checkings.}
}
\value{
an empty plot.
}
\description{
display an empty ggplot2 plot with a text in the middle of the window (for instance to specify that no plot can be drawn).
}
\examples{
### simple example.
gg_empty(text = "NO GRAPH")    
### white page.
gg_empty()
### all the arguments.
gg_empty(text = "NO GRAPH", text.size = 8, title = "GRAPH1", title.size = 10, lib.path = NULL)   
}
