% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_get_legend.R
\name{gg_get_legend}
\alias{gg_get_legend}
\title{gg_get_legend}
\usage{
gg_get_legend(ggplot_built, fun.name = NULL, lib.path = NULL)
}
\arguments{
\item{ggplot_built}{a ggplot build object.}

\item{fun.name}{single character string indicating the name of the function using gg_get_legend() for warning and error messages. Ignored if NULL.}

\item{lib.path}{character vector specifying the absolute pathways of the directories containing the required packages if not in the default directories. Ignored if NULL.}
}
\value{
a list of class c("gtable", "gTree", "grob", "gDesc"), providing legend information of ggplot_built objet, or NULL if the ggplot_built object has no legend.
}
\description{
get legend of ggplot objects #from https://stackoverflow.com/questions/12539348/ggplot-separate-legend-and-plot
}
\examples{
# Simple example
obs1 <- data.frame(time = 1:20, group = rep(c("CLASS_1", "CLASS_2"), times = 10), stringsAsFactors = TRUE) ; p <- ggplot2::ggplot() + ggplot2::geom_point(data = obs1, mapping = ggplot2::aes(x = group, y = time, fill = group)) ; gg_get_legend(ggplot_built = ggplot2::ggplot_build(p))  
#Error message because no legend in the ggplot
obs1 <- data.frame(time = 1:20, group = rep(c("CLASS_1", "CLASS_2"), times = 10), stringsAsFactors = TRUE) ; p <- ggplot2::ggplot() + ggplot2::geom_point(data = obs1, mapping = ggplot2::aes(x = group, y = time)) ; gg_get_legend(ggplot_built = ggplot2::ggplot_build(p))  
}
