% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_point_rast.R
\name{gg_point_rast}
\alias{gg_point_rast}
\title{gg_point_rast}
\usage{
gg_point_rast(
  data = NULL,
  mapping = NULL,
  stat = "identity",
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  raster.width = NULL,
  raster.height = NULL,
  raster.dpi = 300,
  inactivate = TRUE,
  lib.path = NULL
)
}
\arguments{
\item{raster.width}{width of the result image (in inches). Default: deterined by the current device parameters.}

\item{raster.height}{height of the result image (in inches). Default: deterined by the current device parameters.}

\item{raster.dpi}{resolution of the result image.}

\item{inactivate}{logical. Inactivate the fun.name argument of the saferDev::arg_check() function? If TRUE, the name of the saferDev::arg_check() function in error messages coming from this function. Use TRUE if gg_point_rast() is used like this: eval(parse(text = "gg_point_rast")).}

\item{#classical}{arguments of geom_point(), shown here https://rdrr.io/github/VPetukhov/ggrastr/man/geom_point_rast.html}

\item{lib.path:}{character vector specifying the absolute pathways of the directories containing the required packages if not in the default directories. Ignored if NULL.}
}
\value{
a raster scatter plot.
}
\description{
equivalent to ggplot2::geom_point() but in raster mode use it like ggplot2::geom_point() with the main raster.dpi additional argument.
}
\details{
can be long to generate the plot.
use a square plot region. Otherwise, the dots will have ellipsoid shape.
solve the transparency problems with some GUI.
this function is derived from the geom_point_rast() function, created by Viktor Petukhov , and present in the ggrastr package (https://rdrr.io/github/VPetukhov/ggrastr/src/R/geom-point-rast.R, MIT License, Copyright (c) 2017 Viktor Petukhov). Has been placed here to minimize package dependencies.
}
\examples{
# Two pdf in the current directory
set.seed(1) ; data1 = data.frame(x = rnorm(100000), y = rnorm(100000), stringsAsFactors = TRUE) ; saferGraph::open2(pdf.name = "Raster") ; ggplot2::ggplot() + gg_point_rast(data = data1, mapping = ggplot2::aes(x = x, y = y), inactivate = FALSE) ; saferGraph::open2(pdf.name = "Vectorial") ; ggplot2::ggplot() + ggplot2::geom_point(data = data1, mapping = ggplot2::aes(x = x, y = y)) ; dev.off();dev.off() 
}
