% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpalette.R
\name{ggpalette}
\alias{ggpalette}
\title{ggpalette}
\usage{
ggpalette(n, kind = "std", safer_check = TRUE)
}
\arguments{
\item{n}{number of groups on the graph.}

\item{kind}{either "std" for standard gg colors, "dark" for darkened gg colors, or "light" for pastel gg colors.}

\item{safer_check}{Single logical value. Perform some "safer" checks (see https://github.com/safer-r)? If TRUE, checkings are performed before main code running: 1) R classical operators (like "<-") not overwritten by another package because of the R scope and 2) required functions and related packages effectively present in local R lybraries. Set to FALSE if this fonction is used inside another "safer" function to avoid pointless multiple checkings.}
}
\value{
the vector of hexadecimal colors.
}
\description{
provide colors used by ggplot2
the interest is to use another single color that is not the red one used by default
for ggplot2 specifications, see: https://ggplot2.tidyverse.org/articles/ggplot2-specs.html
}
\examples{
ggpalette(n = 2) # output of the function
   
plot(1:7, pch = 16, cex = 5, col = ggpalette(n = 7)) # the ggplot2 palette when asking for 7 different colors
  
plot(1:7, pch = 16, cex = 5, col = ggpalette(n = 7)[5]) # selection of the 5th color of the ggplot2 palette made of 7 different colors
   
plot(1:7, pch = 16, cex = 5, col = ggpalette(n = 7, kind = "dark")) # the ggplot2 palette made of 7 darkened colors
   
plot(1:7, pch = 16, cex = 5, col = ggpalette(n = 7, kind = "light")) # the ggplot2 palette made of 7 lighten colors
}
